src/test/java/com/yaoyao/sell/service/impl/OrderServiceImplTest.java
package com.yaoyao.sell.service.impl;

import com.yaoyao.sell.dataobject.OrderDetail;
import com.yaoyao.sell.dto.OrderDTO;
import lombok.extern.slf4j.Slf4j;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import java.util.ArrayList;
import java.util.List;

import static org.junit.Assert.*;

@RunWith(SpringRunner.class)
@SpringBootTest
@Slf4j
public class OrderServiceImplTest {

    @Autowired
    private OrderServiceImpl orderService;

    private final String BUYEROPENID = "2018517";

    @Test
    public void create() {
        OrderDTO orderDTO = new OrderDTO();
        orderDTO.setBuyerAddress("长沙八骏");
        orderDTO.setBuyerName("凡事都");
        orderDTO.setBuyerPhone("13077718881");
        orderDTO.setBuyerOpenid(BUYEROPENID);
        //购物车
        List<OrderDetail> orderDetailList = new ArrayList<>();
        OrderDetail orderDetail = new OrderDetail();
        orderDetail.setProductId("0001");
        orderDetail.setProductQuantity(1);
        orderDetailList.add(orderDetail);
        orderDTO.setOrderDetailList(orderDetailList);

        OrderDTO result = orderService.create(orderDTO);

        log.info("[create order =====>  ]  ===",result);
    }

    @Test
    public void findOne() {
        OrderDTO orderDTO = orderService.findOne("1526546698603623256");
        log.info("==============",orderDTO);
        Assert.assertEquals("1526546698603623256",orderDTO.getOrderId());
    }

    @Test
    public void findList() {
    }

    @Test
    public void cancel() {
    }

    @Test
    public void finish() {
    }

    @Test
    public void paid() {
    }
}